Add Simple On-boarding Window
Adding a simple on-boarding window with the same blank content as the
main inbox window.

The window currently has close and minimize buttons, but they don't work
and will only look correct on non-Mac systems. Styling and functionality
to come later.

Also adds Aphrodite, which we'll use for styling, and the Source Code
Pro font which we'll use for monospace fonts as needed.