Refactor Menubar Buttons + Font Awesome
Refactoring the menubar buttons to inherit from a base class.

Also added Font Awesome icons and used those for the buttons. The icons
used are pro icons, which is fine while I'm the only dev, but may be an
issue in the future so I'll need to figure out what to do about that.

Test plan:
Run app and make sure minimize and close buttons still work and display
the correct title.