Add Guess Configuration UI
Fleshing out the actual guess configuration UI so that it calls out to
the main thread to guess the configuration of the email address provided

This change is large as it add redux-thunk, and does some code reorg
related to that change. And just general code org improvements too for
good measure.

The screen is mostly still non-functional since it will need a next step
and the main thread actually has to do the guessing, but once we've got
a few more pieces, we should be pretty solid on the account onboarding.