Use Raw Socket for IMAP Test Connections
The IMAP library is heavy and doesn't give us the control we'd like for
testing for IMAP support. Which is fine, it's not supposed to. But we
just want to know if there is a server and, if there is, what auth
capabilities it supports.

To that end, we only need a simple socket connection to accomplish this
so we're switching over to that.

Also removes `user` from a debug log statement. We don't need the PII