:lipstick: Stop Loading CSS Via Unused Import
Prior to this change, we were loading our shared CSS file in a gross
way. We would import it in one of our TS files and rely on injection to
be done when we accessed the file. Which is bonkers.

Now, we do so much cleaner. We just include the base.css file in the
html file. Like it should be. No more weird "we have to import this here
so we get CSS" non-sense.