:sweat_smile: Turn Off "Transpile Only" Typescript
We were doing transpilation only during our webpack builds, which was
convenient when we were getting things in place but basically eliminated
all usefulness from Typescript and it's type checking outside of VS Code
which is not what we want.

So this change turns "transpile only" off. And to do so, it also fixes
every existing type error in the system. Which was a lot. Like a lot a
lot. But that's why we fixed it. And we will hopefully notice them much
easier in the future now that we have this on.

Functionally, nothing should change. These are all type-safety changes.