:key: Type Check Storage Keys
Enable type checking on key/value pairs for each of the storage types.

When defining the types for a storage type, you now need to pass in an
interface with all the keys and their types. This allows more stable
checking and means we don't have to cast types everywhere.

Already caught one bug: `completedOnboarding` -> `completedOnBoarding`