:postal_horn: Add Notifier Class
Similar to our `Router` we now have a `Notifier` class to normalize the
sending of notifications back and forth between processes. Just as with
the `Router`, we also have a list of allowed notification channels and
what is expected by the function being notified.

Note that we don't send `window.minimize` through the notifier since
that is done as a part of browser window initialization, and we don't
want to have to attach the notifer at that point to avoid a dependency
loop.

Also, renamed `window-minimize` to `window.minimize`, just because I
like it better and fits more for future things